Try jzlib to avoid longs in adler32 and crc32 Implemented Deflater using jzlib

+++ jtransc-rt-core/src/com/jtransc/compression/jzlib/Checksum.java
@@ -42,9 +42,9 @@ interface Checksum {
 
 	void reset();
 
-	void reset(long init);
+	void reset(int init);
 
-	long getValue();
+	int getValue();
 
 	Checksum copy();
 }
